Renaming assetAquarius
